Fix: MOP/Cata packet handling
* BATTLEGROUND_NUM_TYPES is raised to 657, we used it in for loops with an uint_8t. This change was implemented in #731
* Note: While we try to loop to 657, the uint8_t stucks at 255... Before changing multiversion stuff, please test your changes :-)

+++ src/world/Management/Battleground/BattlegroundMgr.cpp
@@ -344,7 +344,7 @@ void CBattlegroundManager::HandleGetBattlegroundQueueCommand(WorldSession* m_ses
 
     bool foundSomething = false;
 
-    for (uint8 i = 0; i < BATTLEGROUND_NUM_TYPES; ++i)
+    for (uint32_t i = 0; i < BATTLEGROUND_NUM_TYPES; ++i)
     {
         for (uint8 j = 0; j < MAX_LEVEL_GROUP; ++j)
         {
@@ -585,7 +585,7 @@ void CBattlegroundManager::EventQueueUpdate(bool forceStart)
     m_queueLock.Acquire();
     m_instanceLock.Acquire();
 
-    for (uint8 i = 0; i < BATTLEGROUND_NUM_TYPES; ++i)
+    for (uint32_t i = 0; i < BATTLEGROUND_NUM_TYPES; ++i)
     {
         for (uint8 j = 0; j < MAX_LEVEL_GROUP; ++j)
         {
